% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StudyInvariant.R
\name{loadOhdsiStudy}
\alias{loadOhdsiStudy}
\title{Load OHSDI study}
\usage{
loadOhdsiStudy(file = getDefaultStudyFileName(), verbose = FALSE)
}
\arguments{
\item{file}{(Optional) Name of local file to place results; makre sure to use forward slashes (/)}

\item{verbose}{Logical: print R object names that are loaded}
}
\value{
A list of class type \code{OhsiStudy} that contains all saved study objects
}
\description{
Load OHSDI study
}
\details{
This function loads an OHDSI study results from disk file.
}

