% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StudyInvariant.R
\name{saveOhdsiStudy}
\alias{saveOhdsiStudy}
\title{Save OHDSI study}
\usage{
saveOhdsiStudy(list, file = getDefaultStudyFileName(), compress = "xz",
  includeMetadata = TRUE)
}
\arguments{
\item{list}{A list of R objects to save to disk file.}

\item{file}{(Optional) Name of local file to place results; makre sure to use forward slashes (/)}

\item{compress}{Logical or character string specifying the use of compression. See \code{\link{save}}}

\item{includeMetadata}{Logical: include metadata about user and system in saved file}
}
\description{
Save OHDSI study
}
\details{
This function saves an OHDSI study to disk file.  All objects are written using \code{\link{save}}
format and can be read back from file at a later time by using the function \code{\link{loadOhdsiStudy}}.
}

